Added examples - so many examples. Reported by Vaughn Seward: Describe problem with the /Ref/Api/Functions/Data Type/Array Functions/a5_array_from_string Function.xml_html documentation page:
The syntax is showing as:

N a5_array_from_string(C array_name ,C string ,P vars )

The 3rd parameter is "vars". What is this used for. Is it optional? An example of how to use this would be great.

Or is there another function we should use that can take a string (read in from a file), and convert that to an array of strings?

I found this reference to this function here:

http://www.learnalpha.com/FormObjectUtility/FormObjectUtility.html

The code passes "local_variables()"... not sure what this is for.