Added note for maximum supported index size. Reported by John Poronsky:
Describe problem with the /Desktop/Guide/Design/Data/DBF/Index Builder.xml_html documentation page:

Does not show the maximum index length…my original field was 255 characters…could not index on that because of the length.

I “think” the max length of an index is 100 characters…can you update the index help fields to show max length as you show the maximum length of the index name?